+++ lib/kancha/components/Screen/Screen.tsx
@@ -20,7 +20,7 @@
 import * as React from 'react'
 import { SafeAreaView, ScrollView, ViewStyle, ImageBackground, ImageSourcePropType, StatusBar } from 'react-native'
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scrollview'
-import { Container, Theme, Device } from '@kancha'
+import { Container, Theme, Device, Text } from '@kancha'
 
 /** Spacer size */
 const SPACER_SIZE = 500
@@ -83,6 +83,11 @@ interface ScreenProps {
   footerNavComponent?: React.ReactNode
 
   /**
+   * A footer component that works with KAV
+   */
+  footerNavDivider?: boolean
+
+  /**
    * Use a keyboard avoiding view for text inputs
    */
   keyboardAvoiding?: boolean
@@ -94,7 +99,8 @@ const Screen: React.FunctionComponent<ScreenProps> & {
 } = props => {
   const scrollViewStyle: ViewStyle = {
     backgroundColor: props.type && Theme.colors[props.type].background,
-    // ...(props.footerNavComponent ? { paddingBottom: 100 } : {}),
+    // borderWidth: 1,
+    // borderColor: 'red',
   }
 
   const safeAreaViewStyle = {
@@ -108,39 +114,48 @@ const Screen: React.FunctionComponent<ScreenProps> & {
    * Main content to be rendered
    */
   const mainContent = (
-    <Container background={props.type} flex={1}>
-      <StatusBar hidden={props.statusBarHidden} />
-      {props.children}
-    </Container>
+    <React.Fragment>
+      <Container background={props.type} flex={1}>
+        <StatusBar hidden={props.statusBarHidden} animated />
+        {props.children}
+      </Container>
+    </React.Fragment>
   )
   /**
    * Main content to be rendered within a ScrollView
    */
   const scrollViewContent = (
-    <KeyboardAwareScrollView style={scrollViewStyle}>
-      {props.expandingHeaderContent && (
-        <React.Fragment>
-          <Container backgroundColor={props.headerBackgroundColor}>{props.expandingHeaderContent}</Container>
-          {/* Background image -> <ImageBackground source={Images.backgrounds.purpleGradient}>{props.expandingHeaderContent}</ImageBackground> */}
-        </React.Fragment>
+    <React.Fragment>
+      <KeyboardAwareScrollView style={scrollViewStyle} contentInsetAdjustmentBehavior={'never'}>
+        {props.expandingHeaderContent && (
+          <React.Fragment>
+            <Container backgroundColor={props.headerBackgroundColor}>{props.expandingHeaderContent}</Container>
+          </React.Fragment>
+        )}
+        {mainContent}
+      </KeyboardAwareScrollView>
+
+      {props.footerNavComponent && (
+        <SafeAreaView style={{ backgroundColor: Theme.colors.primary.background }}>
+          <Container paddingTop dividerTop={props.footerNavDivider}>
+            {props.footerNavComponent}
+          </Container>
+        </SafeAreaView>
       )}
-      {mainContent}
-    </KeyboardAwareScrollView>
+    </React.Fragment>
   )
   /**
    * Main content to be rendered within a SafeAreaView
    */
   const safeAreaView = (
     <SafeAreaView style={safeAreaViewStyle}>
-      <Container flex={1}>
-        {props.config === ScreenConfigs.SafeNoScroll ? mainContent : scrollViewContent}
+      <Container flex={1}>{props.config === ScreenConfigs.SafeNoScroll ? mainContent : scrollViewContent}</Container>
 
-        {props.footerNavComponent && (
-          <Container paddingBottom paddingTop>
-            {props.footerNavComponent}
-          </Container>
-        )}
-      </Container>
+      {props.footerNavComponent && props.config === ScreenConfigs.SafeNoScroll && (
+        <Container paddingTop dividerTop={props.footerNavDivider}>
+          {props.footerNavComponent}
+        </Container>
+      )}
     </SafeAreaView>
   )
 
@@ -151,8 +166,7 @@ const Screen: React.FunctionComponent<ScreenProps> & {
     <ImageBackground
       style={{ flex: 1 }}
       source={props.backgroundImage ? props.backgroundImage : {}}
-      resizeMode={'cover'}
-    >
+      resizeMode={'cover'}>
       {safeAreaView}
     </ImageBackground>
   )