+++ lib/kancha/index.ts
@@ -3,6 +3,13 @@
  */
 export { default as Device } from './utilities/device'
 export { default as Strings } from './utilities/strings'
+export { default as Domain } from './utilities/domain'
+export { default as Issuer } from './utilities/issuer'
+
+/**
+ * Services
+ */
+export { default as Toaster } from './services/toast'
 
 /**
  * Theme & Constants
@@ -26,3 +33,10 @@ export { default as NavBar } from './components/NavBar/NavBar'
 export { default as Input } from './components/Input/Input'
 export { default as Scanner } from './components/Scanner/Scanner'
 export { default as ScannerControl } from './components/ScannerControl/ScannerControl'
+export { default as Credential } from './components/Credential/Credential'
+export { default as Toast } from './components/Toast/Toast'
+
+export { default as Request } from './components/Request/Request'
+export { default as Banner } from './components/Banner/Banner'
+export { default as RequestContent } from './components/RequestContent/RequestContent'
+export { default as IndicatorBar } from './components/IndicatorBar/IndicatorBar'