+++ lib/kancha/utilities/domain.ts
@@ -1,0 +1,9 @@
+import URL from 'url-parse'
+
+function domain(url: string) {
+  if (!url) return ''
+  const parsed = new URL(url)
+  return parsed.hostname
+}
+
+export default domain