+++ lib/screens/Screens.ts
@@ -32,6 +32,7 @@ export default {
   KeyChain: 'KeyChain',
 
   Request: 'Request',
+  StaticRequest: 'StaticRequest',
   NewRequest: 'NewRequest',
   NestedInfo: 'NestedInfo',
 
@@ -56,4 +57,13 @@ export default {
     Legacy: 'MigrationLegacy',
     Complete: 'MigrationComplete',
   },
+
+  /**
+   * For testing stuff
+   */
+  Dummy: 'Dummy',
+}
+
+export const COMPONENTS = {
+  Toast: 'Toast',
 }